requirements.txt
Django==3.1.4
asgiref==3.3.1
pytz==2020.5
sqlparse==0.4.1

pytest==6.2.1
attrs==20.3.0
iniconfig==1.1.1
packaging==20.8
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
toml==0.10.2

pytest-django==4.1.0

djangorestframework==3.12.2

psycopg2-binary==2.8.6

drf-yasg==1.20.0
MarkupSafe==1.1.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
jinja2==2.11.2
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
uritemplate==3.0.1
urllib3==1.26.2

django-extensions==3.1.0

parameterized==0.7.4
